# docs/swagger.yaml
basePath: /v1/api
definitions:
  auth.LoginRequest:
    properties:
      email:
        example: exampleuser@mail.com
        type: string
      password:
        example: password
        maxLength: 50
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: exampleuser@mail.com
        type: string
      password:
        example: password
        maxLength: 50
        minLength: 8
        type: string
      username:
        example: exampleuser
        type: string
    required:
    - email
    - password
    - username
    type: object
  book.CreateBookRequest:
    properties:
      author:
        example: F. Scott Fitzgerald
        type: string
      isbn:
        example: "9780743273565"
        type: string
      published_year:
        example: 1925
        type: integer
      title:
        example: The Great Gatsby
        type: string
    required:
    - author
    - isbn
    - published_year
    - title
    type: object
  book.CreateBookResponse:
    properties:
      author:
        example: F. Scott Fitzgerald
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      isbn:
        example: "9780743273565"
        type: string
      published_year:
        example: 1925
        type: integer
      title:
        example: The Great Gatsby
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
    type: object
  book.GetBookResponse:
    properties:
      author:
        example: F. Scott Fitzgerald
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      isbn:
        example: "9780743273565"
        type: string
      published_year:
        example: 1925
        type: integer
      title:
        example: The Great Gatsby
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
    type: object
  book.UpdateBookRequest:
    properties:
      author:
        example: F. Scott Fitzgerald
        type: string
      isbn:
        example: "9780743273565"
        type: string
      published_year:
        example: 1925
        type: integer
      title:
        example: The Great Gatsby
        type: string
    required:
    - author
    - isbn
    - published_year
    - title
    type: object
  book.UpdateBookResponse:
    properties:
      author:
        example: F. Scott Fitzgerald
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      isbn:
        example: "9780743273565"
        type: string
      published_year:
        example: 1925
        type: integer
      title:
        example: The Great Gatsby
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
    type: object
  user.UpdateUserRequest:
    properties:
      email:
        example: exampleuser@mail.com
        type: string
      username:
        example: exampleuser
        type: string
    required:
    - email
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server Book Review server.
  title: Book Review API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login to the application with the provided email and password
      parameters:
      - description: Login details
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Login to the application
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided details
      parameters:
      - description: Registration details
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
      summary: Register a new user
      tags:
      - auth
  /books:
    post:
      consumes:
      - application/json
      description: Create a new book with the provided details. Requires authentication.
      parameters:
      - description: Book details
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/book.CreateBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created book
          schema:
            $ref: '#/definitions/book.CreateBookResponse'
      security:
      - Bearer: []
      summary: Create a new book
      tags:
      - books
  /books/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a book by the provided ID. Requires authentication.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - Bearer: []
      summary: Delete a book by ID
      tags:
      - books
    get:
      consumes:
      - application/json
      description: Get a book by the provided ID. Requires authentication.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/book.GetBookResponse'
      security:
      - Bearer: []
      summary: Get a book by ID
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Update a book with the provided details. Requires authentication.
      parameters:
      - description: Book ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Book details
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/book.UpdateBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/book.UpdateBookResponse'
      security:
      - Bearer: []
      summary: Update a book by ID
      tags:
      - books
  /user/profile:
    delete:
      consumes:
      - application/json
      description: Delete the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - Bearer: []
      summary: Delete user profile
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get user profile
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update the profile of the authenticated user
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - Bearer: []
      summary: Update user profile
      tags:
      - user
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"